+++ packages/mcp/mcp/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @transcend-io/mcp
 
+## 0.6.4
+
+### Patch Changes
+
+- Updated dependencies [212568a]
+  - @transcend-io/mcp-server-docs@0.3.2
+  - @transcend-io/mcp-server-consent@0.4.5
+  - @transcend-io/mcp-server-admin@0.4.4
+  - @transcend-io/mcp-server-assessment@0.4.4
+  - @transcend-io/mcp-server-discovery@0.4.4
+  - @transcend-io/mcp-server-dsr@0.4.4
+  - @transcend-io/mcp-server-inventory@0.4.4
+  - @transcend-io/mcp-server-preferences@0.4.4
+  - @transcend-io/mcp-server-workflows@0.4.4
+
+## 0.6.3
+
+### Patch Changes
+
+- cbe9d3a: Update links in the readmes
+- Updated dependencies [cbe9d3a]
+  - @transcend-io/mcp-server-preferences@0.4.3
+  - @transcend-io/mcp-server-assessment@0.4.3
+  - @transcend-io/mcp-server-discovery@0.4.3
+  - @transcend-io/mcp-server-inventory@0.4.3
+  - @transcend-io/mcp-server-workflows@0.4.3
+  - @transcend-io/mcp-server-consent@0.4.4
+  - @transcend-io/mcp-server-admin@0.4.3
+  - @transcend-io/mcp-server-base@0.6.1
+  - @transcend-io/mcp-server-dsr@0.4.3
+  - @transcend-io/mcp-server-docs@0.3.1
+
 ## 0.6.2
 
 ### Patch Changes